Rename module task -> taskstore
