+++ lib/domain/libimagtodo/src/lib.rs
@@ -47,5 +47,5 @@ module_entry_path_mod!("todo");
 
 pub mod error;
 pub mod result;
-pub mod task;
+pub mod taskstore;
